fix: restore config-driven warmup, dtype kwarg, and lazy boto3 import
The merge auto-resolved SFTConfig in favour of main's side, silently
dropping the branch's warmup_ratio/logging_dir fixes. Restore both and
wire warmup_ratio through from config so the lora/training blocks are
no longer decorative.

- trainer: warmup_ratio from config (was hardcoded warmup_steps=10),
  logging_dir param instead of the TENSORBOARD_LOGGING_DIR env var
- trainer/model_utils: dtype= instead of the deprecated torch_dtype=,
  preserving the fix from db9bba9 that the branch predated
- pipeline: honour --config for the MLflow artifact instead of
  hardcoding config.yaml
- cloud: import boto3 lazily so the module loads without AWS extras
- drop unused imports (sys, Optional, Field, Path, Dict, torch)

+++ litigpt/model_utils.py
@@ -65,7 +65,7 @@ def load_model_and_tokenizer(
         base_model,
         quantization_config=bnb_config,
         device_map="auto",
-        torch_dtype=compute_dtype,
+        dtype=compute_dtype,
         trust_remote_code=True,
     )
